framework commit

+++ internal/model/entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 type User struct {
-	ID           int64  `gorm:"column:id;primaryKey;comment:用户id" json:"id"`                        // 用户id
+	ID           int    `gorm:"column:id;primaryKey;comment:用户id" json:"id"`                        // 用户id
 	UserName     string `gorm:"column:user_name;not null;comment:用户名不允许重复" json:"user_name"`        // 用户名不允许重复
 	AvatarURL    string `gorm:"column:avatar_url;comment:用户头像" json:"avatar_url"`                   // 用户头像
 	UserPassword string `gorm:"column:user_password;not null;comment:用户密码" json:"user_password"`    // 用户密码
